Window and D3DContext classes added
A window class for managing the creation of a window and its events. A
D3DContext class for managing initializing Direct3D11 and cleaning it up.
Also added r2tk as a utility library, with exception handling and
singleton. Added a small demo in the main file.

D3DContext can not yet handle resizing and switching to fullscreen
properly. But as with all DirectX applications, fullscreen mode can be
entered/exited with ALT + ENTER.